chg:move typings into their own directory
