+++ index.ts
@@ -1,3 +1,2 @@
 import { ElementCollection, PageObject, ComponentObject } from "./src";
-
-export default { ElementCollection, PageObject, ComponentObject };
+export { ElementCollection, PageObject, ComponentObject };
