+++ src/types/element.collection.ts
@@ -1,0 +1,17 @@
+import ElementCollection from "src/element.collection";
+
+export type BaseContainerFunction = () => Cypress.Chainable<Cypress.JQueryWithSelector>;
+
+export type Elements = {
+    [key: string]: (...params: any) => Cypress.Chainable<Cypress.JQueryWithSelector>;
+};
+
+export type ComponentObjectFunction = (instance: ElementCollection) => void;
+
+export type NestedComponents = {
+    [key: string]: (fn: ComponentObjectFunction, ...params: any) => void;
+};
+
+export interface IMetadata {
+    [key: string]: any;
+}
