+++ src/types/page.object.ts
@@ -1,0 +1,7 @@
+export interface IPageMetadata {
+    baseUrl: string;
+    path: string | "";
+    title?: string;
+
+    [key: string]: any;
+}
